+++ app/models/project.rb
@@ -1,5 +1,5 @@
+# Contains the structure and elements for a project.
 class Project < ActiveRecord::Base
-
   # Uploaders
   mount_uploader :agency_logo, ImageUploader
   mount_uploader :client_logo, ImageUploader
@@ -12,40 +12,45 @@ class Project < ActiveRecord::Base
     {
       top_level: 'CREATIVE',
       categories: [
-        ["Concepts", "concepts"],
-        ["Treatment", "treatment"],
-        ["Script", "script"],
-        ["Boards", "boards"],
-        ["References", "references"]
+        ['Concepts', 'concepts'],
+        ['Treatment', 'treatment'],
+        ['Script', 'script'],
+        ['Boards', 'boards'],
+        ['Styleframes', 'styleframes'],
+        ['Animatic', 'animatic']
       ]
     },
     {
       top_level: 'TIMELINE',
       categories: [
-        ["Production Calendar", "production-calendar"],
-        ["Agency Overview", "agency-overview"]
+        ['Production Calendar', 'production-calendar'],
+        ['Agency Overview', 'agency-overview']
       ]
     },
     {
       top_level: 'PRODUCTION',
       categories: [
-        ["Production Book", "production-book"],
-        ["Casting", "casting"],
-        ["Locations", "locations"],
-        ["Elements", "elements"]
+        ['Production Book', 'production-book'],
+        ['Casting', 'casting'],
+        ['Talent', 'talent'],
+        ['Locations', 'locations'],
+        ['Style Swipe', 'style-swipe'],
+        ['Set Design', 'set-design'],
+        ['Voiceover', 'voiceovers'],
+        ['Music', 'music']
       ]
     },
     {
-      top_level: 'EDITORIAL',
+      top_level: 'REFERENCES',
       categories: [
-        ["Music", "music"],
-        ["Rough Cuts", "rough-cuts"],
-        ["Final", "final"]
+        ['Client Documents', 'client-documents'],
+        ['Past Work', 'past-work'],
+        ['Visual Style', 'visual-style']
       ]
-    },
+    }
   ]
 
-  TOP_LEVELS = DEFAULT_CATEGORIES.collect{|hash| hash[:top_level]}
+  TOP_LEVELS = DEFAULT_CATEGORIES.collect { |hash| hash[:top_level] }
 
   # Concerns
   include Deletable
@@ -53,13 +58,13 @@ class Project < ActiveRecord::Base
   has_secure_password
 
   # Named Scopes
-  scope :with_editor, lambda { |*args| where('projects.user_id = ? or projects.id in (select project_users.project_id from project_users where project_users.user_id = ? and project_users.editor IN (?))', args.first, args.first, args[1] ).references(:project_users) }
+  scope :with_editor, -> (*args) { where('projects.user_id = ? or projects.id in (select project_users.project_id from project_users where project_users.user_id = ? and project_users.editor IN (?))', args.first, args.first, args[1]).references(:project_users) }
 
   # Model Validation
-  validates_presence_of :name, :user_id
-  validates_uniqueness_of :slug, scope: [ :deleted ], allow_blank: true
-  validates_format_of :slug, with: /\A[a-z][a-z0-9\-]*\Z/, allow_blank: true
-  validates_uniqueness_of :username, :number, scope: [ :deleted ], allow_blank: true
+  validates :name, :user_id, presence: true
+  validates :slug, uniqueness: { scope: :deleted }, allow_blank: true
+  validates :slug, format: { with: /\A[a-z][a-z0-9\-]*\Z/ }, allow_blank: true
+  validates :username, :number, uniqueness: { scope: :deleted }, allow_blank: true
 
   # Model Relationships
   belongs_to :user
@@ -69,7 +74,7 @@ class Project < ActiveRecord::Base
   has_many :locations, -> { where(deleted: false).order(:archived, created_at: :desc) }
   has_many :location_photos
   has_many :project_users
-  has_many :users, -> { where( deleted: false ).order( 'last_name, first_name' ) }, through: :project_users
+  has_many :users, -> { where(deleted: false).order(:last_name, :first_name) }, through: :project_users
   has_many :editors, -> { where('project_users.editor = ? and users.deleted = ?', true, false) }, through: :project_users, source: :user
   has_many :viewers, -> { where('project_users.editor = ? and users.deleted = ?', false, false) }, through: :project_users, source: :user
 
@@ -78,13 +83,13 @@ class Project < ActiveRecord::Base
   end
 
   def self.find_by_param(input)
-    self.where("projects.slug = ? or projects.id = ?", input.to_s, input.to_i).first
+    where('projects.slug = ? or projects.id = ?', input.to_s, input.to_i).first
   end
 
   def grouped_categories_for_select
     groups = []
     TOP_LEVELS.each do |top_level|
-      groups << [top_level, self.categories.where(top_level: top_level).pluck(:name, :id)]
+      groups << [top_level, categories.where(top_level: top_level).pluck(:name, :id)]
     end
     groups
   end
@@ -92,25 +97,29 @@ class Project < ActiveRecord::Base
   # Project Owners and Project Editors
   def editable_by?(current_user)
     @editable_by ||= begin
-      current_user.all_projects.where(id: self.id).count == 1
+      current_user.all_projects.where(id: id).count == 1
     end
   end
 
   # Project Owners
   def deletable_by?(current_user)
-    current_user.projects.where( id: self.id ).count == 1
+    current_user.projects.where(id: id).count == 1
   end
-
 
   private
 
-    def create_default_categories
-      DEFAULT_CATEGORIES.each do |hash|
-        top_level = hash[:top_level]
-        hash[:categories].each_with_index do |(name, slug), index|
-          self.categories.create(top_level: top_level, name: name, slug: slug, position: index+1, user_id: self.user_id)
-        end
+  def create_default_categories
+    DEFAULT_CATEGORIES.each do |hash|
+      top_level = hash[:top_level]
+      hash[:categories].each_with_index do |(name, slug), index|
+        categories.create(
+          top_level: top_level,
+          name: name,
+          slug: slug,
+          position: index + 1,
+          user_id: user_id
+        )
       end
     end
-
+  end
 end