+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.6.0
 
 ### Enhancements
+- **Project Changes**
+  - Top level EDITORIAL has been replaced by REFERENCES
+  - Default categories for projects have been updated
 - **Gem Changes**
   - Updated to turbolinks 2.5.3